+++ src/main/java/dev/willyelton/crystal_tools/client/events/KeyPressEvent.java
@@ -71,7 +71,7 @@ public class KeyPressEvent {
 
         Level level = player.level();
 
-        Levelable levelable = stack.getCapability(Capabilities.ITEM_SKILL, level);
+        Levelable levelable = stack.getCapability(Capabilities.ITEM_SKILL, level.registryAccess());
 
         if (levelable != null) {
             ModGUIs.openScreen(new UpgradeScreen(stack, player, levelable));