+++ src/main/java/dev/willyelton/crystal_tools/client/gui/CrystalBackpackScreen.java
@@ -118,7 +118,7 @@ public class CrystalBackpackScreen extends ScrollableContainerScreen<CrystalBack
         this.addRenderableWidget(new BackpackScreenButton(this.leftPos - 21, screenButtonY, Component.literal("Open Skill Tree"),
                 button -> {
                     this.onClose();
-                    Levelable levelable = this.container.getBackpackStack().getCapability(Capabilities.ITEM_SKILL, player.level());
+                    Levelable levelable = this.container.getBackpackStack().getCapability(Capabilities.ITEM_SKILL, player.level().registryAccess());
                     if (levelable != null) {
                         ModGUIs.openScreen(new UpgradeScreen(menu.getSlotIndex(), menu.getPlayer(), () -> PacketDistributor.sendToServer(new OpenBackpackPayload(menu.getSlotIndex())), levelable));
                     }