+++ src/main/java/dev/willyelton/crystal_tools/common/crafting/CrystalToolsRecipe.java
@@ -1,7 +1,13 @@
 package dev.willyelton.crystal_tools.common.crafting;
 
+import dev.willyelton.crystal_tools.common.capability.Capabilities;
+import dev.willyelton.crystal_tools.common.capability.Levelable;
 import dev.willyelton.crystal_tools.common.components.DataComponents;
+import dev.willyelton.crystal_tools.common.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.common.levelable.skill.SkillPoints;
+import dev.willyelton.crystal_tools.utils.ToolUtils;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.RegistryAccess;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.CraftingBookCategory;
@@ -41,5 +47,21 @@ public abstract class CrystalToolsRecipe extends CustomRecipe {
 
     protected List<Ingredient> getIngredients() {
         return getInputs().stream().map(ItemStack::getItem).map(Ingredient::of).toList();
+    }
+
+    protected void increaseLevelCap(ItemStack stack, HolderLookup.Provider provider, int points) {
+        if (provider instanceof RegistryAccess registryAccess) {
+            Levelable levelable = stack.getCapability(Capabilities.ITEM_SKILL, registryAccess);
+
+            if (levelable != null) {
+                levelable.increaseExpCap(points);
+                return;
+            }
+        }
+
+        // Fallback if for some reason we don't have a real registry access?
+        int newCap = ToolUtils.getNewCap(CrystalToolsConfig.BASE_EXPERIENCE_CAP.get(), points);
+
+        stack.set(DataComponents.EXPERIENCE_CAP, newCap);
     }
 }