+++ changelog.md
@@ -1,7 +1,11 @@
 # Crystal Tools Change Log
 ## 3.0.0 - beta 5
+### Additions
 - Adds new server config option: `vanilla_skill_trees`
   - Allows you to turn on and off vanilla skill trees without a datamap or data pack
+
+### Fixes
+- Fixes issue with the recipes for the crystal elytra and crystal AIOT
 
 ## 3.0.0 - beta 4
 - Updates to Minecraft 1.21.6